add equalize histogram

+++ include/common/types.hpp
@@ -179,8 +179,8 @@ namespace qlm
 	struct Histogram_t 
 	{
 		static constexpr int num_channels = (frmt == ImageFormat::GRAY) ? 1 : 3;
-		static constexpr size_t tot_elements = std::numeric_limits<T>::max() - std::numeric_limits<T>::lowest() + 1;
-
+		static constexpr size_t tot_elements = (size_t)std::numeric_limits<T>::max() - std::numeric_limits<T>::lowest() + 1;
+		// todo : change to pixel
 		std::array<std::array<size_t, tot_elements>, num_channels> hist;
 
 		// Constructor initializes the histograms to zero
